+++ src/library/core/indexers.py
@@ -5,9 +5,8 @@ from library.core.commentextender import ICommentExtenderFields
 @indexer(IComment)
 def comment_picture_indexer(obj):
     # Récupérer l'objet étendu contenant le champ picture    
-    print("indexer")
     extender = ICommentExtenderFields(obj, None)
-    import pdb;pdb.set_trace()
+    # import pdb;pdb.set_trace()
     if extender and extender.picture:        
         return extender.picture
     return None